Create `jpro-internal-util` module and move `PlatformUtils` class into it
