+++ internal/util/build.gradle
@@ -4,11 +4,9 @@ java {
 }
 
 dependencies {
-    implementation project(":internal:util")
     implementation "org.jetbrains:annotations:$JETBRAINS_ANNOTATIONS_VERSION"
     implementation "org.slf4j:slf4j-api:$SLF4J_API_VERSION"
 
-    testImplementation("org.mockito:mockito-core:$MOCKITO_VERSION")
-    testImplementation("org.mockito:mockito-junit-jupiter:$MOCKITO_VERSION")
+    testImplementation "org.assertj:assertj-core:$ASSERTJ_VERSION"
     testRuntimeOnly "ch.qos.logback:logback-classic:$LOGBACK_VERSION"
 }
