+++ internal/util/src/main/java/module-info.java
@@ -1,0 +1,11 @@
+/**
+ * The module descriptor for the internal Process module.
+ *
+ * @author Besmir Beqiri
+ */
+module one.jpro.platform.internal.util {
+    requires org.jetbrains.annotations;
+    requires org.slf4j;
+
+    exports one.jpro.platform.internal.util;
+}
