+++ settings.gradle
@@ -8,7 +8,6 @@ include "jpro-mdfx"
 include "jpro-mdfx:example"
 include "jpro-media"
 include "jpro-media:example"
-include "internal:openlink"
 include "jpro-routing:core"
 include "jpro-routing:core-test"
 include "jpro-routing:dev"
@@ -25,6 +24,8 @@ include "jpro-webrtc"
 include "jpro-webrtc:example"
 include "jpro-sipjs"
 include "jpro-sipjs:example"
+include "internal:openlink"
+include "internal:util"
 include "tree-showing"
 include "freeze-detector"
 include "example"
